refactor(rpc): carry the read-plan result raw after the envelope
The frame gains a second length word, and the bytes it counts follow the envelope untouched by protobuf.
Every RPC but the read-plan reply leaves that part empty, and the read-plan arm of the Response is an empty marker message.
common/rpc_frame.h states the two-word layout and the flat result's magic once for both sides, and each receiver refuses a part past the bound it names.
The build scripts copy common/ next to the plugin sources the way they copy the proto.

+++ proxy/rpc_trace.cc
@@ -103,7 +103,7 @@ void TxRpcTrace::on_stmt(const std::string& sql) {
 }
 
 void TxRpcTrace::record(RpcOp type, uint64_t us, uint32_t req_b,
-                        uint32_t resp_b, const std::string& meta) {
+                        uint64_t resp_b, const std::string& meta) {
   if (!active_) return;
   const uint64_t off = std::chrono::duration_cast<std::chrono::microseconds>(
                            std::chrono::steady_clock::now() - started_)
